requirements.txt
tweepy>=4.14.0
twitchAPI>=4.0.0
python-dotenv>=1.0.0
